+++ frontend/ios/DroidKaigi 2019/data/SessionRepository.swift
@@ -8,7 +8,6 @@
 import Foundation
 import ios_combined
 import RxSwift
-import RxCocoa
 
 final class SessionRepository {
 
@@ -43,39 +42,4 @@ private func handledKotlinException(_ error: Error) -> Error {
     }
 
     fatalError("Unexpedeted KotlinThrowable: \(cause)")
-}
-
-
-// MARK: - Extensions for Kotlin Deferred with RxSwift
-
-extension KotlinThrowable: LocalizedError {
-    public var errorDescription: String? {
-        return self.message ?? "No message. \(self)"
-    }
-}
-
-
-extension Kotlinx_coroutines_core_nativeDeferred {
-    
-    func asSingle<ElementType>(_ elementType: ElementType.Type) -> Single<ElementType> {
-        return Single<ElementType>.create { observer in
-            self.invokeOnCompletion { cause in
-                if let cause = cause {
-                    observer(.error(cause))
-                    return KotlinUnit()
-                }
-
-                if let result = self.getCompleted() as? ElementType {
-                    observer(.success(result))
-                    return KotlinUnit()
-                }
-
-                fatalError("Illegal state or invalid elementType.")
-            }
-
-            return Disposables.create {
-                self.cancel()
-            }
-        }
-    }
 }
